Lang-graph-main/LangGraph/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.10
pydantic>=2.0.0
python-dotenv>=1.0.0
tavily-python>=0.1.9
requests>=2.31.0
groq>=0.4.0
langchain-core>=0.1.0 